+++ central/administration/events/handler/singleton.go
@@ -2,7 +2,7 @@ package handler
 
 import (
 	"github.com/stackrox/rox/central/administration/events/datastore"
-	"github.com/stackrox/rox/pkg/administration/events"
+	"github.com/stackrox/rox/pkg/administration/events/streams"
 	"github.com/stackrox/rox/pkg/sync"
 )
 
@@ -15,7 +15,7 @@ var (
 // Singleton returns an instance of the administration events handler.
 func Singleton() Handler {
 	once.Do(func() {
-		h = newHandler(datastore.Singleton(), events.Singleton())
+		h = newHandler(datastore.Singleton(), streams.Singleton())
 	})
 	return h
 }
